backend/requirements.txt
Django>=5.0.0,<6.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.3.0
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
whitenoise>=6.6.0
python-dotenv>=1.0.0
dj-database-url>=2.1.0
Pillow>=10.0.0 
requests>=2.31.0 